fix: No props in client components when { ssr: false } (#647)
When using `render(..., { ssr: false })`, we wrap client components in order for their initial render result on mount to be `null`, so that we can match the hydrated RSC payload.

When doing this wrapping, we were not passing along the props passed to the wrapper.

As a result, client components used in server components did not get the props passed to them from those server components.

Fixes #583