✨ Added a favicon to the minimal and standard starter (#728)
### Before
Before, we didn't have a favicon at all

<img width="281" height="51" alt="CleanShot 2025-09-13 at 16 45 49" src="https://github.com/user-attachments/assets/870d4399-4475-4cf7-80fb-9dcfad3e82b2" />

### After

This PR adds a light favicon
<img width="293" height="53" alt="CleanShot 2025-09-13 at 16 43 46" src="https://github.com/user-attachments/assets/f8f30a81-0784-44ea-af92-3e1fb5537901" />

and a dark favicon
<img width="293" height="53" alt="CleanShot 2025-09-13 at 16 44 03" src="https://github.com/user-attachments/assets/dc61cfcb-ea04-4313-b230-4264ed6a7db4" />

This will conditionally display, depending on the user's system preferences

### Additional Considerations

This favicon utilizes green and teal colors. The reason that it's not the "standard" brand colors is because our the RWSDK site and our docs both utilize reds and oranges:

<img width="284" height="80" alt="CleanShot 2025-09-13 at 16 44 57" src="https://github.com/user-attachments/assets/bc261130-b75d-40a4-8ed6-cc8058858fe8" />

So, if the user has all 3 of the sites open during development, it can become difficult to visually differentiate these sites.