tests: Add RSC kitchen sink playground example (#769)
This change introduces a new e2e/playground example, `rsc-kitchen-sink`, to demonstrate the framework's React Server Components (RSC) capabilities and cover it with e2e tests. The example includes a server component that renders a client component, which in turn executes two types of server actions: a form action and an `onClick` action. 

We already have our smoke tests covering this, but having it as a playground example helps for debugging purposes when working on the sdk.