feat: Add manual client module overrides + Chakra UI playground` (#777)
### Context

The primary goal was to create a playground to test and demonstrate the integration of the Chakra UI component library with our React Server Components framework, including comprehensive end-to-end tests.

### Hurdles

Several issues were encountered during the implementation.

**1. `"use strict"` Directive Interference**

Solved in #762 (see PR description for context)

**2. Non-Component Exports from Client Modules**

Solved in #774 (see PR description for context)

**3. Server-Side Execution of Client APIs in Chakra UI**

After resolving the framework-level issues, a final blocker emerged. Certain files associated with Chakra UI's `<Code>` component use `createContext`, a React API that is not available in the server environment. These files do not include a `"use client"` directive, which leads to a runtime error.

The files are:
-   [`code-block-context.ts`](https://github.com/chakra-ui/chakra-ui/blob/79971c0d1ccac7921e5e5c65faa93e3fe8456bca/packages/react/src/components/code-block/code-block-adapter-context.ts)
-   [`code-block-adapter-context.ts`](https://github.com/chakra-ui/chakra-ui/blob/79971c0d1ccac7921e5e5c65faa93e3fe8456bca/packages/react/src/components/code-block/code-block-adapter-provider.tsx)

Looking at modules performing the same function for other Chakra UI components, they use "use client" directives. That, combined with the `createContext` call, leads me to believe this was unintentional and should in fact include a `"use client"` directive (I will ask Chakra to get clarity here).

For now, to address this and provide a general-purpose solution for similar cases in other libraries, this PR introduces a `forceClientPaths` option to the RedwoodSDK Vite plugin. This allows developers to manually designate modules as client-side.